deploy/ai/requirements.txt
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

transformers>=4.36.0
albumentations>=2.0.8
opencv-python>=4.8
Pillow>=10.0

numpy>=2.0,<3
pandas>=2.0
scikit-learn>=1.5
tqdm>=4.60
matplotlib>=3.8

fastapi>=0.128.5
uvicorn[standard]>=0.39.0
pydantic>=2.12.5
python-multipart>=0.0.20
httpx>=0.28.1

pytest>=8.0
pytest-cov>=5.0

ruff>=0.6
black>=24.0
mypy>=1.10

pre-commit>=3.7
pip-tools>=7.0